Inital work on moving to rtos wrappers for threads
