+++ STM32F405RGTx/Core/Inc/coulomb_counting_thread.h
@@ -1,0 +1,13 @@
+#include "util.h"
+
+class CoulombCountingThread {
+    public:
+        CoulombCountingThread();
+
+        float getCharge();
+    private:
+        RTOSThread thread;
+        void runCoulombCounting(void* args);
+
+        float prev_current;
+};
